:recycle: Move LanguageSelector to shared widgets directory
This commit relocates the `LanguageSelector` widget to a shared directory for better project organization and reusability.

-   **File Relocation**: Moved `language_selector.dart` from `lib/ui/chat/` to `lib/ui/widgets/`.
-   **Import Updates**: Updated import paths in `ai_chatbox.dart` and `support_page.dart` to reference the new location of `LanguageSelector`.
